Correction Main : prend les nombres decimaux et non les Double

+++ Main.java
@@ -4,8 +4,8 @@ public class Main {
   public static class REPL {
     private List<Operation> operations;
     private Stack<String> pile;
-    private Map<Operation,BiFunction<Double,Double,Double>> interpretations;
-    public <T> REPL(List<Operation> op,List<BiFunction<Double,Double,Double>> applications)
+    private Map<Operation,BiFunction<Integer,Integer,Integer>> interpretations;
+    public <T> REPL(List<Operation> op,List<BiFunction<Integer,Integer,Integer>> applications)
     {
       if (op.size() != applications.size())
       {
@@ -13,18 +13,20 @@ public class Main {
       }
       this.operations = new ArrayList<Operation>(op);
       this.pile = new Stack<String>();
-      this.interpretations = new HashMap<Operation,BiFunction<Double,Double,Double>>();
+      this.interpretations = new HashMap<Operation,BiFunction<Integer,Integer,Integer>>();
       for(int i = 0; i < operations.size(); i++)
       {
         this.interpretations.put(operations.get(i),applications.get(i));
       }
     }
+    //Boucle principal du programme
     public void boucle()
     {
       Scanner scan = new Scanner(System.in);
       while (true)
       {
         String cmd = scan.next();
+        //On quitte la boucle dès que exit est tape par l'utilisateur
         if (cmd.equals("exit"))
         {
           break;
@@ -41,11 +43,11 @@ public class Main {
             }
             else
             {
-              Double arg1 = Double.parseDouble(pile.get(pile.size()-2));
-              Double arg2 = Double.parseDouble(pile.get(pile.size()-1));
+              Integer arg1 = Integer.parseInt(pile.get(pile.size()-2));
+              Integer arg2 = Integer.parseInt(pile.get(pile.size()-1));
               pile.pop();
               pile.pop();
-              Double result = interpretations.get(op).apply(arg1,arg2);
+              Integer result = interpretations.get(op).apply(arg1,arg2);
               System.out.println(result);
               pile.push(String.valueOf(result));
               //System.out.println(pile.pop()+op.getSymbole()+pile.pop());
@@ -67,10 +69,10 @@ public class Main {
     Operation moins = new Operation(2,"-");
     Operation fois = new Operation(2,"*");
     Operation div = new Operation(2,"/");
-    BiFunction<Double,Double,Double> aplus = (a,b) -> a + b;
-    BiFunction<Double,Double,Double> amoins = (a,b) -> a - b;
-    BiFunction<Double,Double,Double> afois = (a,b) -> a * b;
-    BiFunction<Double,Double,Double> adiv = (a,b) -> a / b;
+    BiFunction<Integer,Integer,Integer> aplus = (a,b) -> a + b;
+    BiFunction<Integer,Integer,Integer> amoins = (a,b) -> a - b;
+    BiFunction<Integer,Integer,Integer> afois = (a,b) -> a * b;
+    BiFunction<Integer,Integer,Integer> adiv = (a,b) -> a / b;
     REPL repl = new REPL(List.of(plus,moins,fois,div),
     List.of(aplus,amoins,afois,adiv));
     repl.boucle();
